Move aid related mixin into the `aid` module
