+++ src/minisites/mixins.py
@@ -13,18 +13,6 @@ class SearchMixin:
         return kwargs
 
 
-class AidEditMixin:
-    """Common code to aid editing views."""
-
-    def get_queryset(self):
-        qs = Aid.objects \
-            .filter(author=self.request.user) \
-            .select_related('perimeter') \
-            .order_by('name')
-        self.queryset = qs
-        return super().get_queryset()
-
-
 class NarrowedFiltersMixin:
     """The minisite feature allows an admin to select the available values for
     some filters such as audiences and categories.
